Fix fallback handling, better tests, remove process.env reference
